requirements.txt
requests
python-dotenv
Flask
pyngrok
watchdog
waitress
keyboard
